backend/requirements.txt
alembic>=1.14,<2.0
fastapi>=0.115,<1.0
pwdlib[argon2]>=0.2,<1.0
PyJWT>=2.9,<3.0
psycopg[binary]>=3.2,<4.0
SQLAlchemy>=2.0,<3.0
uvicorn[standard]>=0.30,<1.0
